Move AsyncEnumerable streaming from SliceEncodingExtensions to AsyncEnumerableExtensions
